common.utils: automatic refactor: docstring
Moves `parse_docstring`, `encode_docstring` and `add_header` to
`lrrbot.docstring`.
